notifier/requirements.txt
functions-framework>=3.8.0
google-cloud-secret-manager>=2.20.0
